+++ README.md
@@ -1,0 +1,26 @@
+# 科学计算器
+
+该项目提供了一个科学计算器库以及命令行界面。
+
+## 安装
+
+```bash
+pip install -e .
+```
+
+## 使用
+
+运行交互式计算器：
+
+```bash
+scicalc
+```
+
+或在 Python 中使用：
+
+```python
+from scicalc import evaluate
+print(evaluate("2 + 2"))
+```
+
+wocaonima
