+++ pyproject.toml
@@ -1,0 +1,15 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "scicalc"
+version = "0.1.0"
+description = "Interactive scientific calculator"
+readme = "README.md"
+requires-python = ">=3.8"
+authors = [{name = "Unknown"}]
+dependencies = []
+
+[project.scripts]
+scicalc = "scicalc.__main__:main"
