+++ src/scicalc/__init__.py
@@ -1,0 +1,5 @@
+"""Scientific calculator package."""
+
+from .core import evaluate
+
+__all__ = ["evaluate"]
