+++ src/scicalc/__main__.py
@@ -1,19 +1,10 @@
-#!/usr/bin/env python3
-"""Interactive scientific calculator."""
-import math
+"""Command-line interface for the scientific calculator."""
 
-# Build a dictionary of allowed names from the math module
-ALLOWED_NAMES = {k: getattr(math, k) for k in dir(math) if not k.startswith("_")}
-ALLOWED_NAMES.update({"abs": abs, "round": round})
-
-
-def evaluate(expression: str) -> float:
-    """Safely evaluate a math expression using only allowed names."""
-    expression = expression.replace("^", "**")  # allow ^ for exponent
-    return eval(expression, {"__builtins__": {}}, ALLOWED_NAMES)
+from .core import evaluate
 
 
 def main() -> None:
+    """Run the interactive calculator REPL."""
     print("Scientific Calculator. Type 'exit' or 'quit' to exit.")
     while True:
         try:
